Move licence app base to related contract
